Wrap booking success search params in suspense
